:sparkles: Initial base setup of Jekyll site
This is the base set of content that migrates my v3 website over into a
Jekyll-based static site. The majority of the actual content is missing,
but overall appearance, look, and feel of my site now lives as a Jekyll
site. Woot! :tada:

Have to give a hat tip to @ct-martin as I used the FOSSRIT website as an
example for building out a significant set of this content. It took a
while to get this all figured out, but I'm excited! Jekyll is fun!

Signed-off-by: Justin W. Flory <git@jwf.io>