:link: Fix several broken internal links
Thanks to the new, nifty Travis CI checks introduced in commit
8c383a414bc2a786eb7dac1ba779d5258742d4d0, I discovered several broken
links on the site. I was able to fix most of them by routing them to
the correct place internally on the site. I gave up on `/v3/irc/` since
it is a directory full of several things.

Anyways, this should make the full contents of my website somewhat
easier to navigate. html-proofer is one of the coolest tools I've ever
used for web development, hands-down. So practical!

Signed-off-by: Justin W. Flory <git@jwf.io>