:robot: Disallow search engines to crawl /v3/ except /v3/music/
Search engines have no business to be poking around in the /v3/
directory, so tell them not to do that. Googlebot allegedly is one of
the only/few search engines that listens to the `Allow:` directive, so
we can let it take a look at the `/v3/music/` page because I like that
page. :smile:

Signed-off-by: Justin W. Flory <git@jwf.io>