:construction: Move old site versions to archive/ directory
Now that there are three old versions of my site being hosted in the
same repository, it is a little annoying to have them in the root
directory. To move them out of the way, they will now live in an
`archive/` directory following the same structure they were in the first
place. I also updated the internal links, so CI will still pass.

Signed-off-by: Justin W. Flory <git@jwf.io>