:bowtie: home: Update my activities and projects
This squashed-down commit contains various miscellaneous changes to my
home page. Mostly it updates my current activities and projects, since a
lot of the info is outdated. I also re-arranged a non-trivial amount of
content on different parts of the page.

:key: home: Remove GPG key section

`keys.fedoraproject.org` is offline and I now use an email service that
provides PGP encryption by default. The key is still valid, but the
necessity to promote it is less.

Signed-off-by: Justin W. Flory <git@jwf.io>

:arrow_down: footer: Move SF Conservancy badge into footer

Signed-off-by: Justin W. Flory <git@jwf.io>

:penguin: home: Update Fedora activities

Move a few lines around, update some links to new addresses.

Signed-off-by: Justin W. Flory <git@jwf.io>

:sassy_man: home: Update "Who am I" section

I did it! Sort of. Almost. Soon enough. :man_student:

Signed-off-by: Justin W. Flory <git@jwf.io>

:baby: home: Update UNICEF activities

Move in the content I already have on my blog:

    https://blog.justinwflory.com/about-me/

Signed-off-by: Justin W. Flory <git@jwf.io>

:telephone: Update TeleIRC activities

TeleIRC is Go now, not just Node. :smile:

Signed-off-by: Justin W. Flory <git@jwf.io>

:earth_americas: home: Add LibreCorps section

Since this is a major aspect of what I do these days, it should probably
go on the website.

Signed-off-by: Justin W. Flory <git@jwf.io>