:tada: Refactor: Manage site content in Markdown with layouts (#11)
This is a non-trivial, behind-the-scenes rebuild of how my site comes
together. Instead of being a Jekyll site that serves static HTML, it is
now a Jekyll site that builds from modular content written in Markdown,
which is then generated into static HTML. Virtually nothing changes in
how the site actually appears, but this is almost a total rewrite and
leverages the more powerful functionality that Jekyll, Front Matter, and
Liquid syntax offers.

This is a really big "Feels Good" moment, because I have wanted to do
something like this for my site for years! Before, I didn't have the
technical knowledge to really understand how to do it. But now, I was
able to do this in a few hours of work. A little bit of a confidence
booster for me. :smile:

This will warrant a new git tag.

Signed-off-by: Justin W. Flory <git@jwf.io>