:beetle: projects: Fix empty string for id attributes
I used the wrong object name for the `id` field. So, when Jekyll went to
build the site, the `id` tag was empty for every project. This also
meant HTML did not validate. This fixes that. I also changed `item` to
`project` since I thought that was more obvious in meaning.

Signed-off-by: Justin W. Flory <git@jwf.io>