:money_with_wings: home: "Hire me!" section for seeking opportunities
This adds a new section with a config option for `hire_me`. When I am
looking for work, I can set this to `true`, and my LinkedIn and contact
info will appear, along with a notice that I am for-hire. It hides my
LinkedIn otherwise, because I prefer not to have people look over it
unless I am in the market.

Signed-off-by: Justin W. Flory <git@jwf.io>