:world_map: :boom: archive: Remove archived pages from sitemap
This was causing S.E.O. issues with Google because these pages were
included in my sitemap despite being blocked by the `robots.txt`. So,
per documentation of `jekyll-sitemap`, this ignores these committed
pages in the sitemap by adding Front Matter to hide them.

I confirmed this fix by building the site locally and checking the
contents of `sitemap.xml`. It is exactly how I want it to be.

Signed-off-by: Justin W. Flory <git@jwf.io>