:beetle: :link: archive: Fix all invalid HTML
Previously I ignored the errors and problems on the old, archived
versions of my past website. But it keeps hanging up the CI pipeline and
I figured, might as well fix it once and get it over with.

So now there is only valid HTML on the whole website. Hurray!

Signed-off-by: Justin W. Flory <git@jwf.io>