:beetle: :link: spigotmc: Drop non-public URL from website
Previously, I linked to the thread I posted in 2015 on the SpigotMC
forums with the story, pictures, and other details about MINECON 2015.
However, the category the thread was posted in was archived and requires
a login to view the thread.

So, it breaks my CI and even if someone did click the link, they would
have to log in to see it. It's a nice idea but I'm dropping this link.

Signed-off-by: Justin W. Flory <git@jwf.io>