:rainbow: meta: Update GitHub refs to GitLab
This commit used the great powers of `grep` to discover all of the
places where I was linking to the old GitHub Pages repo. And this commit
promptly replaces it with the GitLab repo. In most cases, I actually
used a variable for the git repo URL in case it ever changes again
someday.

Signed-off-by: Justin W. Flory <git@jwf.io>