:free: legal: Relicense from MPLv2 to GPLv3
The MPLv2 is the copyleft software license I have used by default since
high school. While I still like the MPLv2 for what it is, I now
understand that it is a bit of an unusual license to use for a static
website made with Jekyll.

So, to be more direct and clear in my intent, I am relicensing the site
under the GNU Public License v3.

Signed-off-by: Justin W. Flory <git@jwf.io>