:dart: nav: Improve responsiveness of navbar for multiple layouts
This builds on the previous change where I migrated to Bootstrap. This
commit leverages some cool Jekyll ways of loading a nav bar in from YAML
data and makes sure that it is fluid and responsive.

This isn't perfect to me, but it is much better than what was here
previously. So, all for gradual improvements! :raised_hands:

Signed-off-by: Justin W. Flory <git@jwf.io>