:page_with_curl: legal: Initial commit
First commit of a legal notice I am adding to my website. This legal
notice creates a policy and precedent for default licensing for digital
contributions I make on the Internet.

This is an attempt to better protect and safeguard my creations and work
under a license acceptable to my uses.

Signed-off-by: Justin W. Flory <git@jwf.io>