:wrench: index: Condense projects to index.adoc
This commit removes the content/projects.html include file and moves it
into the index page. Since this content is only used once, yet it is
dynamically created as an include, does not make great sense.

I believe this makes the index page easier to maintain by having more
things in fewer places. This is more useful to me than the several
include partials that I was only using because I did not fully
understand Jekyll then.

Signed-off-by: Justin W. Flory <git@jwf.io>