:new: :purple_heart: federation: Initial commit
This is an initial commit of an idea I have sat with for most of 2020.
This is the first time I have sat down and really let myself explore the
idea more fully.

For now, I don't have a lot of additional work to do with the Federation
of Intersectional Communities, but it will serve as a grounding point
for my forthcoming endeavors. Setting this up unblocks those other
endeavors.

I'm excited for this. This is something I really believe in… at least,
right now. :grinning:

Signed-off-by: Justin W. Flory <git@jwf.io>