:memo: fedora: Update CommOps and Docs Team
This commit updates my Fedora bio info. It moves CommOps back to my
active areas of focus in Fedora, because truthfully… I never really left
CommOps. I also added Fedora Docs because I do a decent amount of work
there and it is an active area of focus for me.

Signed-off-by: Justin W. Flory <git@jwf.io>