:mag: :link: blog: Use site.blog directive always over static links
This commit is a benefit to the S.E.O. for my personal blog. I am trying
to phase out the old domain (`justinwflory.com`) and get my new one
(`jwf.io`) propagated in the search engines.

Who knows if I will reap any benefit from this in a week, or six years.
:man_shrugging:

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>