:wrench: :beetle: ci: Ignore intentionally broken URLs
The URL that existed here before was never real to begin with; it
pointed to a domain I no longer control. So, I tried to future-proof
myself in case I ever find the original content by using a URL for
somewhere I actually own.

TL;DNR: Was always broken, but doing it this way, maybe one day it will
not be.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>