:thought_balloon: README: Remove Travis badge, clarify legal intent
I no longer rely on Travis CI for this repository, so don't show a
meaningless badge to mislead.

I clarified my legal intent behind "or any later version" for the GPL. I
only intend for "any later version" written and published by the Free
Software Foundation, based in Massachusetts, United States. In the event
that the FSF were to fold and another organization published an update
to the GPL, this project shall not be relicensed.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>