:two: :zero: :two: :one: config: s/2020/2021
This commit updates the year. We are now in 2021… so the footer may as
well reflect that. Not that it really matters anyways, but I think it is
cute to have the year reflected in the footer. It implies someone is
actually looking at it. :laughing:

Also shortened some of my site description. I am a minimalist, what can
I say?

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>