:penguin: :envelope: 2021 life update: Retire TeleIRC, update Fedora
This commit is a periodic update on life for me, at least where I am
today. While I retain commit access to TeleIRC, I will keep it up on my
website. But I no longer actively work on it like I did when I was a
university student, so I am putting it into past-tense.

I also updated various parts of the Fedora Project part of my biography.
I also raised it higher on the page in order to give it more prominence,
because I really do a lot in Fedora even still.

AND OF COURSE NEW FEDORA LOGO AHHHHH.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>