:new: :free: legal: Change project license from GPL-3.0 to MPL-2.0
Given all the license chit-chat I do in my dayjob, I have come to
recognize the GPL-3.0 is not an effective license choice for how I wish
this site to be governed. Instead, the MPL-2.0 is a better fit for this,
especially as I work towards migrating to Hugo and breaking this site
into a more modular theme that could be repurposed and reused in other
creative works.

All work created under this repository is hereto licensed under the
Mozilla Public License 2.0. :tada:

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>