:city_sunrise: Migrate site from Jekyll to Hugo
This commit is a complete, initial migration from the v4 Jekyll-powered
site to a new v5 Hugo-powered site. Across this commit, the existing
content, theme, and styling is shifted from Jekyll into Hugo. It is
deliberately done to avoid major changes to look-and-feel.

Going forward, there are new opportunities to leverage mechanics and
features of Hugo to build a more modular site theme, and increase
simplicity of maintenance. The main purpose of this exercise is to build
my own comfort in working with Hugo, to apply the knowledge learned to
my workings with the UNICEF Open Source Inventory.

Some highlights and key changes of this Pull Request are below:

* v4 Jekyll site is now archived at `/archive/v4/`
* Create a new Hugo theme, titled "Toph", for my site styling
* Translate Jekyll config values to Hugo config values
	* NOTE: Content cannot contain values from configs without creating
	  custom shortcodes. I did not create custom shortcodes to interface
	  with the config values, so many values across the site revert to
	  hard-coding text or URLs. This should be corrected in the future.

Furthermore, the raw HTML output is not clean or correct at this time. A
lot of the `seo-meta.html` values need to be cleaned properly or fixed
so that search engines will understand them correctly. This is very much
a work-in-progress!

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

:alembic: ci: Migrate pipeline from Jekyll to Hugo

To correspond with the previous commit, this commit reworks the GitLab
CI pipeline to work with Hugo and HTMLProofer. It uses the
`cibuilds/hugo` container image maintained by the CircleCI team, as I
use for GitHub Pages. I added the following flags to the HTMLProofer
check that were not implemented before:

* `--check-favicon`
* `--check-img-http`
* `--enforce-https`

The `yaml_lint` and `hugo_build` jobs on GitLab are confirmed to pass
successfully as of this commit.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>