:bookmark_tabs: content: Keep root-level pages outside of a category
The `federation.adoc` and `legal.adoc` files are meant to be standalone
pages, not a part of a category. Putting them in `content/pages/` was
confusing and actually not necessary. It also created extraneous content
for the sitemaps by putting them in a category, and then overriding the
URL. So, this fixes that and reduces some complexity.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>