:package: :arrow_up: theme: Update Bootstrap from 4.5.2 to 5.1.3
This commit represents a notable shift in how my website is built and
served. The underlying web development framework, Bootstrap, is taken
from v4 to v5. This drops jQuery as a dependency, which means pulling in
less JavaScript and faster page loads.

This reworks some parts of the site to fit v5, but more notably, it
actually gets more of the site **using** Bootstrap classes in the first
place. Previously, a lot of my old website styling from v2/v3 were still
lingering in the form of my custom CSS. Note that this commit greatly
reduces the amount of custom CSS maintained.

Some stylistic changes were made, such as a change in default fonts for
headers and the site text, and adding background colors to the project
section `<h3>` elements.

Built and tested locally with Hugo v0.89.1.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>