:briefcase: layouts/partials: Add new for-hire.html partial
This commit separates the "Hire me" page logic from the index file and
moves it into its own partial file. This partial is now embedded inside
the same logic block in the index.html file. The code is removed from
the index file.

Some other refactors and content edits were made to the default content
on the index page.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>