:new: :carousel_horse: layouts/partials: Add carousel of project logos
This commit adds a new feature to the home page. A carousel of all the
project logos will rotate towards the top of the page. Each logo is
linked to the corresponding section further down on the page. This is
a handy improvement to navigation since my home page may fit into
several viewports.

It opens the possibility in the future for each of these sections to
exist as a caption in the carousel, but I'll sit with this change for a
while before making a new change.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>