:map: :ghost: theme: Add boolean for hiding pages from sitemap
This commit creates a custom sitemap template in the Hugo theme. There
is now a check whether a page has defined the `hide_sitemap` variable in
the front matter. If this is set to `true`, the page will be omitted
from the sitemap, so it is not indexed by search engines.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>