:footprints: :link: themes: Convert footer to template-driven layout
This commit changes the theme to use a template-driven layout for the
footer instead of hard-coding specific links and badges into the theme
footer. This work moves the site theme closer to a more general-purpose
use case that others could use.

It also creates three new footer badges in the form of AsciiDoc files,
which are used as the source content for the footer content.

The theme README is also updated with documentation about this feature.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>