:dart: theme: De-duplicate single page template from base template
Previously, the single page template was overriding and duplicating the
base template in `_default/baseof.html`. This meant the same code would
need to be updated in multiple places if making changes. It converts the
single page template to define a `main` block, which in turn is rendered
within the base template.

Less complexity, no visual changes to the front-end. Everyone wins.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>