:globe_with_meridians: theme: Add internationalization support
This commit reworks the theme layouts to support internationalization.
The theme now has an English localization file bundled by default with
the theme. This makes it easier to localize the Hugo theme into other
languages as needed.

Only English is supported as of now, but this may change to include
Spanish and Hindi in the future, to give my translation skills a
challenge. :wink:

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>