🆕🇪🇸 i18n: Add Spanish translations
First attempt to provide greater internationalization support for the
Toph theme, using my own personal site as a testing example. For now,
only Spanish translations are available, although I hope to expand this
to Arabic soon as well. That said, I am also building the site (or
trying to) with respect to language direction, i.e. left-to-right or
right-to-left languages.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>