:lipstick: theme: Vertically center text in project carousel
Projects without a logo use the title element in the project carousel.
The title element is now vertically centered in the project carousel
without changing the height of the whole carousel after a transition.

This looks good on a desktop view, but the mobile site needs larger text
to be more responsive on smaller viewports.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>