:lipstick: theme: Add styling to table of contents (#toc)
This commit adds styling to the table of contents so it is more visually
distinct on a page. This assumes the default table of contents IDs
created by Hugo when using Asciidoctor TOCs.

An example of this can be found on the "Legal" page.

In the future, this could be collapsible, and perhaps the unordered list
could be made into an ordered list.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>