:new: :wrench: Manage site colors and fonts from Hugo config
This commit enables a site to quickly define custom colors and fonts
from the Hugo configuration, without making changes to the theme
templates. It makes the theme easier to use and customize, without hard-
coding specific colors and fonts into the theme.

This is a contributing effort into releasing a general-purpose theme
from Toph, instead of keeping it to my personal site.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>