:wrench: :bug: favicon: Load from config file or default value
This commit fixes the favicon behavior in the theme. Previously,
`params.paths.favicon` was being set and never used in the theme. Now,
the `head.html` partial correctly loads the favicon from the config file
if it is set at `params.favicon`.

Additionally, if `params.favicon` is not set, a default value of
`favicon.ico` will be used instead. This is also how I get to exclude
the value from my site config file, to improve simplicity. :sunglasses:

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>