:bug: :link: assets: Bundle CSS in assets, do not hard-code stylesheet
This commit makes the theme more resilient, particularly for sites that
may not be served from the root domain. Now, the CSS document is looked
up within the Hugo site, minified, and a relative permalink is used to
embed the stylesheet into the page.

This fixes a bug when I was building an example site served from an
example base domain.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>