:new: :building_construction: theme: Add example site for using theme
This commit adds an example site that can be quickly adapted for someone
else who wishes to use this theme in a new Hugo site. To demonstrate the
look and feel of the Hugo theme, presenting my own research and writings
on Septima Clark.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

:construction_worker: exampleSite: Add CI builds for example site

Does not store or do anything with the HTML for now.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>