:1234: projects: Change file order using weighted file names
Any time I add a new entry, I rename every file in order to preserve the
file order that matches how Hugo will parse the profiles using the
`date` field. This change to the numbering sequence preserves the
general priority of how I want the information to be ordered to a user.

The numbering system I improvised was to this logic:

* 0x: Current, active engagements highlighted at the top
* 4x: Other professional associations and prior employment
* 8x: Personal projects and volunteer contributions

New files, obligations, and groupings can be added in the future without
requiring a rename of every file in the repository. This preserves a
git log change history that is easier to review and compare.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>