:green_heart: CI: Ignore HTML-Proofer checks on /spigotmc.org/
Because the Spigot website uses a Cloudflare firewall, HTML-Proofer
always has troubles when verifying that the website is online. The
checks consistently fail even though the website does in fact exist. So,
this is the easiest solution, and if Spigot does go offline, then I can
manually correct that whenever the time may come.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

:green_heart: archive: Fix a broken link redirect

The CI pipeline fails because some pages are inaccessible behind a
firewall. This fixes the CI pipeline by using the direct link instead of
the shortened link.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>