:green_heart: CI: Fix permissions issue in html-tidy job
The artifact retrieved from the previous job has a permissions issue.
The Hugo site is built in the previous step as a user with the UID 3434.
This is a mismatch from the user running in this step, circleci (UID
1001). If the below `chown` step is omitted, the `htmlbeautifer` script
will fail because it is unable to write to the directory.

This behavior changed between October 2022 and March 2023, likely due to
some change in the Hugo container image used in the previous step.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>