⏪️ Revert 1360f75f48b34a5fe4a977821d18b21f2ef3e88a
I thought I knew what I was doing, but when I moved the images out of
the `static/` directory, they were no longer layered onto the root
directory as `/img/`. So, I thought I was being clever, but I was not
using this feature of Hugo correctly, so I am reverting it.

Signed-off-by: Justin W. Wheeler <git@jww.sh>