🌐 content: Add Arabic and Hindi translations
This commit adds a translated version of my biography in Arabic and
Hindi. I have double-checked this translations with Google Translate,
but I would prefer to have a native speaker sign off and review the
translations.

A future idea might be to add a "verified" option, to indicate whether
translations are human-verified or machine-generated.

Signed-off-by: Justin Wheeler <git@jwheel.org>